Explicitly grant START_ACTIVITIES_FROM_BACKGROUND to com.android.phone.
It's needed because when we broadcast ACTION_SIM_SLOT_STATUS_CHANGED
we want to allow the receiving app to start an activity from the
background.

The app already has it implicitly, since it has the same shared UID as
com.android.stk which has the permission for unrelated reasons. Making
it explicit makes it less likely it will lose the permission
accidentally re-introducing a subtly bug.

Bug: 132691768
Test: Builds
Change-Id: Id27fbac478d24d3445007a1437f56867f20acc0f
Merged-In: Id27fbac478d24d3445007a1437f56867f20acc0f
Signed-off-by: rezaadi0105 <rezaadipangestu5@gmail.com>